+++ xenon-common/src/main/java/com/vmware/xenon/common/ServiceRequestListener.java
@@ -16,6 +16,8 @@ package com.vmware.xenon.common;
 import java.io.IOException;
 import java.net.URI;
 
+import io.netty.handler.codec.http.cors.CorsConfig;
+
 public interface ServiceRequestListener {
 
     public static final String PROPERTY_NAME_RESPONSE_PAYLOAD_SIZE =
@@ -52,4 +54,6 @@ public interface ServiceRequestListener {
     void setSecureAuthCookie(boolean secureAuthCookie);
 
     boolean getSecureAuthCookie();
+
+    void setCorsConfig(CorsConfig corsConfig);
 }