+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 1.6.2-SNAPSHOT
+
+* Add CORS support.
+  Using netty's `CorsHandler`, CORS support is added to `ServiceHost`.
+  To configure CORS, override newly added `ServiceHost#getCorsConfig()` method.
+  It is disabled by default.
+
+
 ## 1.6.1
 
 * Fix RootNamespaceService isn't found with local URI GET